fix: start in MainDrawer and add stack.screen(Home) to start in mini games
