+++ screens/MainScreen.js
@@ -15,6 +15,7 @@ import Home from "./Home";
 import Chatting from "./Chatting";
 import Etc from "./Etc";
 import CalendarScreen from "./CalendarScreen";
+import Lab from "./Lab";
 
 const Tab = createBottomTabNavigator();
 
@@ -25,6 +26,7 @@ export default function MainScreen({ navigation }) {
       <Tab.Screen name="Album" component={Album} options={{headerShown:false}}/>
       <Tab.Screen name="Chatting" component={Chatting} options={{headerShown:false}}/>
       <Tab.Screen name="Calendar" component={CalendarScreen} options={{headerShown:false}}/>
+      <Tab.Screen name="Lab" component={Lab} options={{headerShown:false}}/>
     </Tab.Navigator>
   );
 }
